refactor: De-uglify/mess modules
