+++ muridae/muridae/MuridaeWeb/Route/Item.hs
@@ -4,25 +4,20 @@
 {-# LANGUAGE ExplicitNamespaces #-}
 {-# LANGUAGE TypeOperators #-}
 
-module MuridaeWeb.Routes.ItemListings where
+module MuridaeWeb.Route.Item where
 
 import GHC.Generics (Generic)
+import MuridaeWeb.Handler.Item.Types (TradableItem)
+import qualified MuridaeWeb.Route.ItemListing as ItemListings
 import Servant (JSON)
 import Servant.API.Generic (type (:-))
 import Servant.API.NamedRoutes (NamedRoutes)
 import Servant.API.Verbs (Get)
-import MuridaeWeb.Handlers.Items.Listings.Index (TradableItemListing)
-import Servant.API.Capture (Capture)
-import MuridaeWeb.Handlers.Items.Types (TradableItemId)
-import Servant.API (type (:>))
 
 type Routes = NamedRoutes Routes'
 
 data Routes' mode = Routes'
-  { index ::
-      mode
-        :- Capture "item_id" TradableItemId
-        :> "listings"
-        :> Get '[JSON] [TradableItemListing]
+  { indexItems :: mode :- Get '[JSON] [TradableItem]
+  , listings :: mode :- ItemListings.Routes
   }
   deriving stock (Generic)
